feat: env vars for genesis kreation
